requirements.txt
asgiref==3.4.1
dj-database-url==0.5.0
Django==4.0
django-cors-headers==3.10.1
django-heroku==0.3.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
gunicorn==20.1.0
joblib==1.1.0
numpy==1.22.0
onnx==1.10.2
Pillow==8.4.0
protobuf==3.19.1
psycopg2==2.9.3
PyJWT==2.3.0
pytz==2021.3
PyYAML==6.0
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
sqlparse==0.4.2
tensorflow-addons==0.15.0
threadpoolctl==3.0.0
typeguard==2.13.3
typing_extensions==4.0.1
tzdata==2021.5
whitenoise==5.3.0
scikit-learn==1.0.1